fix: fixed weird use before initialization bug
Import cycles, really confusing if you've never seen it before.

[ci skip]
